requirements.txt
accelerate==0.15.0
azure-storage-blob==12.14.1
black
datasets==2.7.1
diffusers==0.10.2
python-dotenv==0.21.0
pillow==9.30
pyarrow==10.0.1
tqdm==4.64.1
transformers==4.25.1
torch==1.13.0
torchvision==0.14.0
wandb==0.13.6